Refactor Product table to use DECIMAL type for monetary values and add categoryId foreign key

+++ migrations/20250331113507-add-Product-Table.js
@@ -28,15 +28,15 @@ export default {
         allowNull: false,
       },
       discount_amt: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10, 2),
         allowNull: false,
       },
       discount_rate:{
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10,2),
         allowNull: false,
       },
       tax_rate: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10,2),
         allowNull: false,
       },
       tax_type: {
